category page test moved to jest (for memory testing purposes) [skip ci]
